Refactor component imports to use consistent casing; add ErrorMessage and SaveDialog components for improved error handling and save functionality
