+++ src/components/categorySelector/index.tsx
