+++ src/components/errorMessage/index.tsx
@@ -3,7 +3,7 @@ interface ErrorMessageProps {
     message: string;
 }
 
-export const ErrorMessage: React.FC<ErrorMessageProps> = ({ message }) => {
+const ErrorMessage: React.FC<ErrorMessageProps> = ({ message }) => {
     return (
         <div className="bg-red-50 border border-red-200 text-red-700 px-4 py-3 rounded relative">
             <strong className="font-bold">Error: </strong>
@@ -11,3 +11,5 @@ export const ErrorMessage: React.FC<ErrorMessageProps> = ({ message }) => {
         </div>
     );
 }
+
+export default ErrorMessage;
