+++ src/components/knowledge/KnowledgeContent.tsx
@@ -1,5 +1,5 @@
 import SharedContent from '@/components/share/SharedContent';
-import { ModelSelector } from "@/components/ui/model-selector";
+import { ModelSelector } from "@/components/ui/modelSelector";
 import { clearCachedAnswers } from '@/store/interview/slice';
 import { SharedItem, SharedCategoryShuffled, User, SavedItem, FollowUpQuestion, ItemTypeSaved } from '@/types/common';
 import { useDispatch } from 'react-redux';
