+++ src/pages/ChatPage.tsx
@@ -3,8 +3,8 @@ import { Layout } from '@/layouts';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { useTranslation } from 'react-i18next';
-import { ModelSelector } from '@/components/ui/model-selector';
-import { SaveDialog } from '@/components/ui/save-dialog';
+import { ModelSelector } from '@/components/ui/modelSelector';
+import { SaveDialog } from '@/components/ui/saveDialog';
 import { MessageItem } from '@/components/chat/message-item';
 import { SendHorizontal, AlertCircle, Sparkles } from 'lucide-react';
 import { Alert, AlertDescription } from '@/components/ui/alert';
